revise PR
