+++ crates/core/Cargo.toml
@@ -7,10 +7,14 @@ license = "MIT"
 include = ["src/lib.rs"]
 
 [dependencies]
-ts-rs = "6.1"
-serde = { version = "1", features = ["derive"] }
-serde-wasm-bindgen = "0.4"
+serde = { version = "1", features = ["derive"], optional = true }
+serde-wasm-bindgen = { version = "0.4", optional = true }
 wasm-bindgen = "0.2.84"
 
+[dev-dependencies]
+ts-rs = "6.1"
+
 [features]
-serde-enable = []
+default = ["serde", "serde-wasm-bindgen"]
+serde-wasm-bindgen = ["dep:serde-wasm-bindgen"]
+serde = ["dep:serde"]
