+++ crates/web/Cargo.toml
@@ -9,8 +9,11 @@ crate-type = ["cdylib"]
 
 [dependencies]
 console_error_panic_hook = "0.1"
-rose = { path = "../core", features = ["serde-enable"] }
-serde = { version = "1", features = ["derive"] }
-serde-wasm-bindgen = "0.4"
+rose = { path = "../core", features = ["serde", "serde-wasm-bindgen"] }
+serde = { version = "1", features = ["derive"], optional = true }
+serde-wasm-bindgen = { version = "0.4", optional = true }
 ts-rs = "6"
 wasm-bindgen = "=0.2.84"
+
+[features]
+default = ["serde", "serde-wasm-bindgen"]
