+++ crates/web/src/lib.rs
@@ -1,5 +1,6 @@
 use rose::Module;
-use wasm_bindgen::prelude::wasm_bindgen;
+use serde::{Deserialize, Serialize};
+use wasm_bindgen::prelude::{wasm_bindgen, JsValue};
 
 #[wasm_bindgen]
 pub fn initialize() {
@@ -7,16 +8,20 @@ pub fn initialize() {
 }
 
 #[wasm_bindgen]
-pub fn greet(name: String) -> String {
-    format!("Hello, {}!", name)
+#[cfg(feature = "serde")]
+pub fn module_from_js(my_mod_js: JsValue) -> Result<Module, JsValue> {
+    // Deserialize the JavaScript object to the Rust Module struct
+    let my_mod: Module = serde_wasm_bindgen::from_value(my_mod_js)?;
+    Ok(my_mod)
 }
 
 #[wasm_bindgen]
-pub fn double(x: u32) -> u32 {
-    x * 2
+#[cfg(feature = "serde")]
+pub fn module_to_js(my_mod: &Module) -> JsValue {
+    // Serialize the modified Module object back to a JsValue and return the modified JsValue
+    to_js_value(&my_mod).unwrap()
 }
 
-#[wasm_bindgen]
-pub fn module(x: Module) -> Module {
-    x
+fn to_js_value(value: &(impl Serialize + ?Sized)) -> Result<JsValue, serde_wasm_bindgen::Error> {
+    value.serialize(&serde_wasm_bindgen::Serializer::json_compatible())
 }
