+++ package.json
@@ -14,9 +14,9 @@
   },
   "scripts": {
     "build": "yarn build-ts && yarn workspace @rose-lang/core build",
-    "build-ts": "rimraf packages/core/src/bindings && cargo test --package=rose && cp -r crates/core/bindings packages/core/src/bindings/ && rimraf crates/core/bindings && yarn prettier --write packages/core/src/bindings/",
+    "build-ts": "rimraf packages/wasm/bindings && cargo test --package=rose && cp -r crates/core/bindings packages/wasm/bindings/ && rimraf crates/core/bindings",
     "vscode": "yarn workspace rose package",
-    "wasm": "cargo build --package=rose-web --target=wasm32-unknown-unknown --release && cargo build --package=rose --features serde-enable --target=wasm32-unknown-unknown --release && yarn workspace @rose-lang/wasm build",
+    "wasm": "cargo build --package=rose-web --target=wasm32-unknown-unknown --release && yarn workspace @rose-lang/wasm build",
     "test": "vitest"
   }
 }
