+++ packages/core/src/translate.test.ts
@@ -1,18 +1,6 @@
-import { double, greet, module_from_js, module_to_js } from "@rose-lang/wasm";
+import { module_from_js, module_to_js } from "@rose-lang/wasm";
+import { Module } from "@rose-lang/wasm/bindings/Module.js";
 import { expect, test } from "vitest";
-import { Module } from "./bindings/Module.js";
-
-test("test Rust Hello World", () => {
-  expect(greet("Raven")).toBe("Hello, Raven!");
-});
-
-test("test different Rust Hello World", () => {
-  expect(double(2)).toBe(4);
-});
-
-test("test Module type", () => {
-  expect("types" in myMod).toBe(true);
-});
 
 let myMod: Module = {
   types: [],
