+++ packages/wasm/.gitignore
@@ -1,2 +1,3 @@
 /dist/
 /wbg/
+/bindings/
